Rename player views
