+++ Demo/Sources/BasicPlaybackView.swift
@@ -1,0 +1,56 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import CoreMedia
+import Player
+import SwiftUI
+
+// MARK: View
+
+// Behavior: h-exp, v-hug
+@available(tvOS, unavailable)
+private struct TimeSlider: View {
+    @ObservedObject var player: Player
+    @StateObject var progressTracker = ProgressTracker(
+        interval: CMTime(value: 1, timescale: 10),
+        seekBehavior: UserDefaults.standard.seekBehavior
+    )
+
+    var body: some View {
+        Slider(progressTracker: progressTracker)
+            .bind(progressTracker, to: player)
+            .debugBodyCounter()
+    }
+}
+
+// Behavior: h-exp, v-exp
+struct BasicPlaybackView: View {
+    @ObservedObject var player: Player
+
+    var body: some View {
+        ZStack {
+            VideoView(player: player)
+#if os(iOS)
+            TimeSlider(player: player)
+                .padding()
+                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .bottom)
+#endif
+            if player.isBuffering {
+                ProgressView()
+            }
+        }
+    }
+}
+
+// MARK: Preview
+
+struct BasicPlaybackView_Previews: PreviewProvider {
+    static var previews: some View {
+        BasicPlaybackView(player: Player())
+            .background(.black)
+            .previewLayout(.fixed(width: 320, height: 180))
+    }
+}
