+++ Demo/Sources/MultiView.swift
@@ -21,8 +21,8 @@ struct MultiView: View {
     var body: some View {
         VStack(spacing: 10) {
             Group {
-                PlaybackView(player: topPlayer)
-                PlaybackView(player: bottomPlayer)
+                BasicPlaybackView(player: topPlayer)
+                BasicPlaybackView(player: bottomPlayer)
             }
             .background(.black)
         }
